Week 4 - New Item
Added a link to the new item page on the home.
Created a new folder for the new page. Aswell as the new item function, only focusing on the quantity of the new item this week. Minimum of 1 up to 20. Able to use the buttons to add quantity or just type it in.